added automatisation of build process
