+++ package.json
@@ -6,13 +6,20 @@
     "@htmlacademy/editorconfig-cli": "1.0.0",
     "autoprefixer": "10.2.5",
     "browser-sync": "2.26.14",
+    "del": "6.0.0",
     "gulp": "4.0.2",
-    "gulp-csso": "^4.0.1",
+    "gulp-htmlmin": "5.0.1",
+    "gulp-imagemin": "7.1.0",
     "gulp-plumber": "1.2.1",
     "gulp-postcss": "9.0.0",
+    "gulp-rename": "2.0.0",
     "gulp-sass": "4.1.0",
     "gulp-sourcemaps": "3.0.0",
+    "gulp-svgstore": "7.0.1",
+    "gulp-terser": "^2.0.1",
+    "gulp-webp": "4.0.1",
     "postcss": "8.2.8",
+    "postcss-csso": "5.0.1",
     "stylelint": "13.12.0",
     "stylelint-config-htmlacademy": "0.1.4"
   },
@@ -20,9 +27,8 @@
     "editorconfig": "editorconfig-cli",
     "stylelint": "stylelint \"source/sass/**/*.scss\" --syntax scss",
     "test": "npm run editorconfig && npm run stylelint",
-    "build": "gulp styles",
-    "start": "gulp",
-    "html": "gulp html"
+    "build": "gulp build",
+    "start": "gulp"
   },
   "browserslist": [
     "last 2 versions",
@@ -39,9 +45,5 @@
   ],
   "engines": {
     "node": "14.15.0"
-  },
-  "dependencies": {
-    "gulp-htmlmin": "^5.0.1",
-    "gulp-rename": "^2.0.0"
   }
 }
