+++ source/companions.html
@@ -4,8 +4,8 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="css/style.css" rel="stylesheet">
-  <link rel="icon" href="favicon.png" sizes="16x16" type="image/png">
+  <link href="../build/css/style.min.css" rel="stylesheet">
+  <link rel="icon" href="../build/favicon.png" sizes="16x16" type="image/png">
   <title>Попутчики</title>
 </head>
 
@@ -1741,9 +1741,7 @@
       </div>
     </div>
   </footer>
-  <script src="js/navigationActions.js"></script>
-  <script src="js/filterActions.js"></script>
-  <script src="js/checkboxeActions.js"></script>
+  <script src="js/script.js"></script>
 </body>
 
 </html>
