+++ source/directions.html
@@ -8,8 +8,8 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="css/style.css" rel="stylesheet">
-  <link rel="icon" href="favicon.png" sizes="16x16" type="image/png">
+  <link href="../build/css/style.min.css" rel="stylesheet">
+  <link rel="icon" href="../build/favicon.png" sizes="16x16" type="image/png">
   <title>Направления</title>
 </head>
 
@@ -1202,8 +1202,7 @@
       </div>
     </div>
   </footer>
-  <script src="js/navigationActions.js"></script>
-  <script src="js/selectCountry.js"></script>
+  <script src="js/script.js"></script>
 </body>
 
 </html>
