+++ source/index.html
@@ -8,8 +8,8 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="css/style.css" rel="stylesheet">
-  <link rel="icon" href="favicon.png" sizes="16x16" type="image/png">
+  <link href="../build/css/style.min.css" rel="stylesheet">
+  <link rel="icon" href="../build/favicon.png" sizes="16x16" type="image/png">
   <title>Погнали!</title>
 </head>
 
@@ -855,8 +855,7 @@
       </div>
     </div>
   </footer>
-  <script src="js/navigationActions.js"></script>
-  <script src="js/tariffs.js"></script>
+  <script src="../build/js/script.min.js"></script>
 </body>
 
 </html>
